backend/pyproject.toml
[project]
name = "todo-ai-chatbot-backend"
version = "0.1.0"
description = "Backend for Todo AI Chatbot with MCP-first architecture, OpenAI Agent SDK, and OpenAI ChatKit integration"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.1",
    "sqlmodel>=0.0.8",
    "httpx>=0.25.0",
    "python-multipart>=0.0.6",
    "better-auth>=0.0.1b11",
    "openai>=1.3.5",
    "openai-agents>=0.2.9",
    "pydantic-settings>=2.1.0",
    "uvicorn>=0.23.2",
    "psycopg2-binary>=2.9.7",
    "alembic>=1.12.0",
    "python-jose>=3.3.0",
    "passlib>=1.7.4",
    "bcrypt>=4.0.1",
    "python-dotenv>=1.0.0",
]
